refactor(#8): move out issue_info from tracker package and get rid of issue_description class ✨
