+++ issue-solver/src/issue_solver/issues/trackers/github_issue_tracker.py
@@ -5,12 +5,11 @@ from pydantic import AnyUrl, Field
 from pydantic_core import Url
 from pydantic_settings import SettingsConfigDict
 
+from issue_solver.issues.issue import IssueReference, IssueInfo
 from issue_solver.issues.trackers.issue_tracker import (
     IssueTracker,
-    IssueInfo,
-    IssueReference,
 )
-from issue_solver.issues.settings import ApiBasedIssueTrackerSettings
+from issue_solver.issues.trackers.settings import ApiBasedIssueTrackerSettings
 
 
 class GithubObjectType(StrEnum):
