+++ issue-solver/src/issue_solver/issues/trackers/gitlab_issue_tracker.py
@@ -8,14 +8,16 @@ from pydantic import AnyUrl, Field
 from pydantic_core import Url
 from pydantic_settings import SettingsConfigDict
 
+from issue_solver.issues.issue import (
+    IssueId,
+    IssueInternalId,
+    IssueReference,
+    IssueInfo,
+)
 from issue_solver.issues.trackers.issue_tracker import (
     IssueTracker,
-    IssueInfo,
-    IssueReference,
-    IssueInternalId,
-    IssueId,
 )
-from issue_solver.issues.settings import ApiBasedIssueTrackerSettings
+from issue_solver.issues.trackers.settings import ApiBasedIssueTrackerSettings
 
 
 class GitlabObjectType(StrEnum):
