+++ issue-solver/src/issue_solver/issues/trackers/http_based_issue_tracker.py
@@ -2,12 +2,11 @@ from typing import Literal
 
 from pydantic_settings import SettingsConfigDict
 
+from issue_solver.issues.issue import IssueReference, IssueInfo
 from issue_solver.issues.trackers.issue_tracker import (
     IssueTracker,
-    IssueInfo,
-    IssueReference,
 )
-from issue_solver.issues.settings import ApiBasedIssueTrackerSettings
+from issue_solver.issues.trackers.settings import ApiBasedIssueTrackerSettings
 
 
 class HttpBasedIssueTracker(IssueTracker):
