+++ issue-solver/src/issue_solver/issues/trackers/issue_tracker.py
@@ -1,26 +1,6 @@
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from pathlib import Path
 
-
-@dataclass
-class IssueInfo:
-    description: str
-    title: str | None = None
-
-
-@dataclass
-class IssueId:
-    id: str
-
-
-@dataclass
-class IssueInternalId:
-    project_id: str
-    iid: str
-
-
-IssueReference = IssueId | IssueInternalId | Path
+from issue_solver.issues.issue import IssueReference, IssueInfo
 
 
 class IssueTracker(ABC):
