+++ issue-solver/src/issue_solver/issues/trackers/settings.py
